Move functions for resizing the MUC sidebar to converse-muc-chatarea
